# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hearsay",
    "description": "Local-first meeting-note transcriber. The core serves this API on `127.0.0.1` only, for the bundled web UI.\n\n**Conventions**\n\n- Base path is `/api`. Every request needs the session token; see the `session_token` scheme.\n- List endpoints return `{ total, page, page_size, items }` and take `page` (at least 1, default 1) and `page_size` (1 to 200, default 50).\n- Errors return `{ \"detail\": \"...\" }`. Database and internal failures render as a literal \"internal error\" — details go to the log only.\n- Times are ISO 8601. Segment `start_s` and `end_s` are seconds relative to the meeting start, not wall clock.\n\n**Not in this document**\n\nThree served routes have no OpenAPI representation and are specified in `docs/api.md`: the live transcript WebSocket at `/ws/meetings/{id}`, meeting audio at `/api/meetings/{id}/audio` (a byte stream with `Range` support), and `/`, which serves the UI with the token injected.",
    "license": {
      "name": "Apache-2.0 WITH Commons-Clause",
      "url": "https://github.com/rwiseman87/hearsay/blob/main/LICENSE"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/identities": {
      "get": {
        "tags": [
          "speakers"
        ],
        "summary": "List known people, most recently updated first.",
        "description": "Every name that has ever been bound to a speaker, whether or not it has a stored voiceprint.\nPowers the rename autocomplete, so a name used in one meeting is suggested in the next. For the\nsubset that can actually be recognized by voice, see `GET /api/voiceprints`.",
        "operationId": "list_identities",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of known people",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_IdentityRead"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/meetings/{id}": {
      "get": {
        "tags": [
          "meetings"
        ],
        "summary": "Read one meeting.",
        "operationId": "get_meeting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The meeting",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingRead"
                }
              }
            }
          },
          "404": {
            "description": "No such meeting"
          }
        }
      },
      "delete": {
        "tags": [
          "meetings"
        ],
        "summary": "Delete a meeting and its recordings.",
        "description": "Removes the database rows (segments, clusters, and both notes tables cascade) and deletes the\non-disk folder. Irreversible. The active recording session cannot be deleted — stop it first.",
        "operationId": "delete_meeting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "No such meeting"
          },
          "409": {
            "description": "The meeting is currently recording"
          }
        }
      },
      "patch": {
        "tags": [
          "meetings"
        ],
        "summary": "Rename a meeting.",
        "description": "Updates the title only; filing a meeting into a folder is a separate operation\n(`PUT /api/meetings/{id}/folder`). The title is trimmed before it is stored.",
        "operationId": "update_meeting",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The renamed meeting",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingRead"
                }
              }
            }
          },
          "404": {
            "description": "No such meeting"
          },
          "422": {
            "description": "Blank title, or over 255 characters"
          }
        }
      }
    },
...
    "/api/meetings/{id}/rediarize": {
      "post": {
        "tags": [
          "speakers"
        ],
        "summary": "Re-run the offline refine (\"Refine speakers\").",
        "description": "Re-diarizes and re-transcribes the whole Them track: global clustering, overlap handling, and\nvoiceprint recognition of people named in earlier meetings. More accurate than the streaming\nlabels, which run online with limited context.\n\nManual (locked) names are carried across. Segment-level work is not: per-line reassignments and\nhand-edited text are rebuilt only at the cluster level, which is what the `edited` flag warns\nabout before the user starts a refine.",
        "operationId": "rediarize",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The meeting's speakers after the refine",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SpeakerRead"
                }
              }
            }
          },
          "404": {
            "description": "No such meeting"
          },
          "409": {
            "description": "The meeting is still recording"
          },
          "503": {
            "description": "The refine model or diarizer sidecar is unavailable"
          }
        }
      }
    },
...
    "/api/meetings/{id}/segments/{segment_id}/speaker": {
      "patch": {
        "tags": [
          "meetings"
        ],
        "summary": "Reassign one line to a different speaker.",
        "description": "Fixes an individual diarization mistake without renaming the whole cluster. Send exactly one of\n`cluster_id` (move the line to an existing speaker in this meeting) or `display_name` (assign it\nto a person by name, reusing that identity's cluster in the meeting if it has one, else creating\na new locked speaker).\n\nThe line is flagged `edited`, so it counts toward the warning shown before a refine discards\nmanual work. The database is the source of truth; `transcript.md` is re-exported best-effort.",
        "operationId": "reassign_segment_speaker",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "segment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SegmentSpeakerAssign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The reassigned segment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentRead"
                }
              }
            }
          },
          "404": {
            "description": "The segment is not in this meeting"
          },
          "409": {
            "description": "The meeting is still recording"
          },
          "422": {
            "description": "A Me line, an unknown target cluster, or a body that is not exactly one of the two fields"
          }
        }
      }
    },
    "/api/meetings/{id}/speakers": {
      "get": {
        "tags": [
          "speakers"
        ],
        "summary": "List a meeting's diarized speakers.",
        "description": "Each item carries the cluster's resolved `label`: a bound name if the speaker has been named,\nelse `Speaker N`. Me is the microphone channel, not a cluster, so it never appears here.",
        "operationId": "list_speakers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of this meeting's speakers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SpeakerRead"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{id}/speakers/{cluster_id}": {
      "put": {
        "tags": [
          "speakers"
        ],
        "summary": "Name a speaker, binding the cluster to a cross-meeting identity.",
        "description": "Gets or creates the identity by name, locks the binding, and relabels that speaker's segments.\nA locked binding wins over voiceprint recognition and survives a re-diarize. If the meeting is\nstill active the name also propagates to the live pipeline, so subsequent utterances carry it.",
        "operationId": "rename_speaker",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeakerRename"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The renamed speaker",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeakerRead"
                }
              }
            }
          },
          "404": {
            "description": "The cluster is not in this meeting"
          },
          "422": {
            "description": "Blank name"
          }
        }
      }
    },
    "/api/meetings/{id}/speakers/{cluster_id}/merge": {
      "post": {
        "tags": [
          "speakers"
        ],
        "summary": "Combine two speakers the diarizer split apart.",
        "description": "Folds the cluster named in the path into the one named by `into`: its lines are repointed,\nrelabelled to the target's name, and flagged `edited`, and the now-empty cluster is deleted.\nScoped to this meeting — no other meeting is touched, and the target's stored voiceprint is left\nalone. The source's own voice sample goes with its row, so a spurious split stops feeding\nrecognition.\n\nThe freed ordinal is not reused, so labels can read \"Speaker 1, Speaker 3\"; renumbering would\nrename unrelated speakers whose lines still carry the old text.",
        "operationId": "merge_speakers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeakerMerge"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The meeting's speakers after the merge",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SpeakerRead"
                }
              }
            }
          },
          "404": {
            "description": "The path cluster is not in this meeting"
          },
          "409": {
            "description": "The meeting is still recording"
          },
          "422": {
            "description": "A self-merge, or an `into` cluster outside this meeting"
          }
        }
      }
    },
...
    "/api/settings/speakers": {
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Replace the speaker-recognition settings.",
        "description": "Full-replaces the section. `recognition_threshold` is the cosine similarity a stored voiceprint\nmust clear to auto-name a returning speaker: higher is stricter.",
        "operationId": "update_speakers",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeakerSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The stored section",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeakerSettings"
                }
              }
            }
          },
          "422": {
            "description": "A missing field, or a recognition threshold outside 0.0..=1.0"
          }
        }
      }
    },
...
    "/api/voiceprints": {
      "get": {
        "tags": [
          "voiceprints"
        ],
        "summary": "The voice roster: people who have a stored voiceprint, with all of their samples.",
        "description": "Lists only people who **have** an embedding. Someone named in a meeting that was never refined\nhas no centroid and is therefore absent — they are still a known person on `GET /api/identities`,\njust not a voice. Clearing someone's last sample drops them from this list, so `sample_count` is\nalways at least 1.\n\n`active_count` is the subset recognition actually uses (a sample counts only once its name was\nset by hand), so `active_count: 0` means \"saved, matching nothing yet\". `dimension` is the\nembedding length — 256 on macOS, 192 on Windows — and samples of different lengths never match\neach other.",
        "operationId": "list_voiceprints",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of people with stored voiceprints",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_VoiceprintRead"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MeetingRead": {
        "type": "object",
        "description": "A meeting row for the API.",
        "required": [
          "id",
          "title",
          "folder",
          "status",
          "started_at",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "ended_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "folder": {
            "type": "string",
            "description": "The on-disk recordings-directory name (not an organizational folder). To move a meeting\nbetween user folders, use `folder_id`."
          },
          "folder_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "The organizational [`FolderRead`] this meeting is filed under, or `null` when unfiled."
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "started_at": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/MeetingStatus"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "MeetingUpdate": {
        "type": "object",
        "description": "Rename a meeting. `title` replaces the meeting's display title (validated non-empty, <= 255 chars).",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string"
          }
        }
      },
...
      "Page_IdentityRead": {
        "type": "object",
        "description": "Paginated list envelope used by every list endpoint (`{ total, page, page_size, items }`).",
        "required": [
          "total",
          "page",
          "page_size",
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "A known cross-meeting person (offered as a rename suggestion).",
              "required": [
                "id",
                "display_name"
              ],
              "properties": {
                "display_name": {
                  "type": "string"
                },
                "email": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "id": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "page_size": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Page_SpeakerRead": {
        "type": "object",
        "description": "Paginated list envelope used by every list endpoint (`{ total, page, page_size, items }`).",
        "required": [
          "total",
          "page",
          "page_size",
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "A diarization cluster within a meeting, with its resolved display label.",
              "required": [
                "id",
                "ordinal",
                "label",
                "locked"
              ],
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "description": "The cluster id."
                },
                "identity_id": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid"
                },
                "label": {
                  "type": "string",
                  "description": "The bound identity's name, else `\"Speaker {ordinal}\"`."
                },
                "locked": {
                  "type": "boolean"
                },
                "ordinal": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "page_size": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "SegmentRead": {
        "type": "object",
        "description": "A transcript segment for the API.",
        "required": [
          "id",
          "stream",
          "speaker_label",
          "text",
          "start_s",
          "end_s",
          "edited"
        ],
        "properties": {
          "cluster_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "edited": {
            "type": "boolean",
            "description": "Whether this segment has been manually edited — its text or its speaker assignment (drives the\n\"edited\" badge + the discard-on-refine warning)."
          },
          "end_s": {
            "type": "number",
            "format": "double"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "speaker_label": {
            "type": "string"
          },
          "start_s": {
            "type": "number",
            "format": "double"
          },
          "stream": {
            "$ref": "#/components/schemas/Stream"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "SegmentSpeakerAssign": {
        "type": "object",
        "description": "Reassign a single Them line to a different speaker. Provide exactly one of: `cluster_id` (an\nexisting cluster in the meeting) or `display_name` (a person by name — reuses that identity's\ncluster in the meeting if it has one, else creates a new locked speaker).",
        "properties": {
          "cluster_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "SpeakerRead": {
        "type": "object",
        "description": "A diarization cluster within a meeting, with its resolved display label.",
        "required": [
          "id",
          "ordinal",
          "label",
          "locked"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The cluster id."
          },
          "identity_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "label": {
            "type": "string",
            "description": "The bound identity's name, else `\"Speaker {ordinal}\"`."
          },
          "locked": {
            "type": "boolean"
          },
          "ordinal": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "SpeakerRename": {
        "type": "object",
        "description": "Rename a cluster to a person (binds + locks; relabels that speaker's segments).",
        "required": [
          "display_name"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          }
        }
      },
...
      "Stream": {
        "type": "string",
        "description": "Which capture channel a segment came from.",
        "enum": [
          "me",
          "them"
        ]
      },
...
    },
    "securitySchemes": {
      "session_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-process session token. `make rust-serve` prints it in the `open:` URL; the packaged app passes it to the webview through a handshake file. Channels that cannot set a header — the WebSocket and the audio element — take `?token=` instead."
      }
    }
  },
  "security": [
    {
      "session_token": []
    }
  ],
...
}